GUI: Use the APPSEL one-shot layer on GUI+W
The goal here is to have a key, that when tapped, sends GUI+W (which
pops ups an app selector with an input field), and turn on the APPSEL
layer that has shortcuts for the most switched to apps. This APPSEL
layer needs to be one-shot, and disappear if any other key is pressed,
or after use, as to not conflict with other things.

This is accomplished by setting the layer on, up in ONESHOT_LAYER state,
and then immediately clearing the PRESSED state of it - thus it
registers as an OSL tap.

This fixes #3.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>