ARRW: Replace Backspace with Enter
This is so that I can control Plex with the arrow layer toggled on. With
this change, most of #83 is addressed, but `Space` is still needed
somewhere, so that stuff can be paused... or I can just switch back to
the default layer once I started the film.

Therefore, this fixes #83.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>