tools: New tool that to listen for commands on the console
This new tool is intended to be chained after `hid_listen`, and will
listen to commands on the HID console, and run custom actions if it
finds any. The intent is to move things that are problematic and
unportable in the firmware, into user-space.

Things that require knowledge of operating system state and such, are
best handled there, and the keyboard only triggers the execution of
these scripts.

Fixes #120.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>